Allow string replacement in the input command.
This in the spirit of xargs. Specifying a string to -I causes that string to
be replaced with the iteration number of the current command. Use:

  multitime -I{} -i "cmd {}" 3 cmdtotest

cmd will be called "cmd 0", "cmd 1", "cmd 2" each time it is asked to
generate inputs. This is useful if you need to pipe different inputs to the
command in each execution.