Add '-o' switch to allow programs output to be tested.
The idea here is twofold. If you just want to record the output, use something
like:
  -o "cat >> file"
or:
  -o "tee file"

If you want to check that the program's output is correct, specify a command
of your choice. It will be fed the output from the main command run by
multitime on stdin. The "output command" can then process this as it wishes:
if its return code is anything but 0, multitime assumes a problem has
occurred, and stops. The idea here is that running a program multiple times
can sometimes expose bugs: merely running things a lot isn't very useful if
you're not sure that the program is doing the right thing.